+++ src/test/operators/jit_operator/operators/jit_read_write_tuple_test.cpp
@@ -318,7 +318,7 @@ TEST_F(JitReadWriteTupleTest, UseValueIDsFromReferenceSegment) {
   JitReadTuples read_tuples;
   bool use_actual_value{false};
   auto a_tuple_entry = read_tuples.add_input_column(DataType::Int, true, ColumnID{0}, use_actual_value);
-  AllTypeVariant value{int64_t{4321}};
+  AllTypeVariant value{int32_t{4321}};
   auto literal_a_tuple_entry = read_tuples.add_literal_value(value);
   auto literal_b_tuple_entry = read_tuples.add_literal_value(value);
